+++ astro.config.mjs
@@ -2,6 +2,7 @@ import { defineConfig } from 'astro/config';
 import mdx from '@astrojs/mdx';
 import sitemap from '@astrojs/sitemap';
 import expressiveCode from 'astro-expressive-code';
+import react from '@astrojs/react';
 
 // https://astro.build/config
 export default defineConfig({
@@ -22,6 +23,7 @@ export default defineConfig({
     }),
     mdx(),
     sitemap(),
+    react(),
   ],
   // output: 'static' is the default — deploys to GitHub Pages
 });